Common: Fixed vape bug

+++ src/PuffSmith/Vape/Dto/PatchDto.php
@@ -87,9 +87,9 @@ class PatchDto extends AbstractDto {
 	 */
 	public ?int $juice;
 	/**
-	 * @var int|null
+	 * @var float|null
 	 */
-	public ?int $power;
+	public ?float $power;
 	/**
 	 * @var int|null
 	 */
