+++ src/PuffSmith/Vape/Dto/VapeDto.php
@@ -107,9 +107,9 @@ class VapeDto extends AbstractDto {
 	 */
 	public ?int $juice;
 	/**
-	 * @var int|null
+	 * @var float|null
 	 */
-	public ?int $power;
+	public ?float $power;
 	/**
 	 * @var int|null
 	 */
